Fixed compile and linker errors

+++ fceu/types.h
@@ -74,7 +74,7 @@ typedef uint32_t uint32;
 #endif
 
 #ifdef __GNUC__
- typedef uint64_t uint64;
+ typedef unsigned long long uint64;
  typedef uint64 u64;
  typedef long long int64;
  #define INLINE inline
